Remove X3 examples
The existence of this "example" directory introduces significant amount
of noise in GitHub full-text search and `grep`. Therefore I decide to
just remove it for now, as we can reference the exsisting code in the
`boostorg/spirit` repository anyway.

Generally speaking, "example" directory is very, very hard to maintain
without having sufficient number of active contributors. We currently
lack the human resources and can't maintain the divergence of the
codebase.

Furthermore, I believe the code snippets should be included in the (new)
documentation, as the documentation itself should be self-contained in
the first place.